refactor(cli): split typer command groups wave 4
