+++ src/cli/commands/settings.py
@@ -15,7 +15,13 @@ import argparse
 import asyncio
 from datetime import datetime, timezone
 
+import typer
+
 from src.cli import runtime
+from src.cli.commands.common import (
+    apply_startup,
+    run_async,
+)
 
 
 async def get_impl(config_path: str, *, key: str | None = None) -> None:
@@ -259,3 +265,129 @@ def run(args: argparse.Namespace) -> None:
                 api_key=getattr(args, "api_key", None),
             )
         )
+
+
+# --------------------------------------------------------------------------- #
+# settings → get / set / info / server-time / agent / filter-criteria
+#            / reactions / semantic
+# --------------------------------------------------------------------------- #
+
+settings_app = typer.Typer(
+    invoke_without_command=True, help="System settings management"
+)
+
+
+@settings_app.callback()
+def settings_main(ctx: typer.Context) -> None:
+    """Bare ``settings`` (no sub-command) runs ``get`` — argparse parity (#1123 review).
+
+    The legacy dispatcher defaulted ``settings_action`` to ``get`` and listed all
+    settings; preserve that on the direct Typer surface, not just the argparse
+    bridge. With a sub-command this is a no-op and the sub-command runs normally.
+    """
+    if ctx.invoked_subcommand is None:
+        apply_startup(ctx)
+        run_async(get_impl(ctx.obj.config, key=None))
+
+
+@settings_app.command("get")
+def settings_get(
+    ctx: typer.Context,
+    key: str | None = typer.Option(None, "--key", help="Specific setting key (default: show all)"),
+) -> None:
+    """Show settings."""
+    apply_startup(ctx)
+    run_async(get_impl(ctx.obj.config, key=key))
+
+
+@settings_app.command("set")
+def settings_set(
+    ctx: typer.Context,
+    key: str = typer.Argument(..., help="Setting key"),
+    value: str = typer.Argument(..., help="Setting value"),
+) -> None:
+    """Set a setting value."""
+    apply_startup(ctx)
+    run_async(set_impl(ctx.obj.config, key=key, value=value))
+
+
+@settings_app.command("info")
+def settings_info(ctx: typer.Context) -> None:
+    """Show system diagnostics."""
+    apply_startup(ctx)
+    run_async(info_impl(ctx.obj.config))
+
+
+@settings_app.command("server-time")
+def settings_server_time(ctx: typer.Context) -> None:
+    """Show current server time (UTC)."""
+    apply_startup(ctx)
+    run_async(server_time_impl(ctx.obj.config))
+
+
+@settings_app.command("agent")
+def settings_agent(
+    ctx: typer.Context,
+    backend: str | None = typer.Option(
+        None, "--backend", help="Agent backend override (auto, claude, deepagents, codex, adk)"
+    ),
+    prompt_template: str | None = typer.Option(
+        None, "--prompt-template", help="Default prompt template"
+    ),
+) -> None:
+    """Configure agent backend and defaults."""
+    apply_startup(ctx)
+    run_async(
+        agent_impl(ctx.obj.config, backend=backend, prompt_template=prompt_template)
+    )
+
+
+@settings_app.command("filter-criteria")
+def settings_filter_criteria(
+    ctx: typer.Context,
+    min_uniqueness: float | None = typer.Option(None, "--min-uniqueness"),
+    min_sub_ratio: float | None = typer.Option(None, "--min-sub-ratio"),
+    max_cross_dupe: float | None = typer.Option(None, "--max-cross-dupe"),
+    min_cyrillic: float | None = typer.Option(None, "--min-cyrillic"),
+) -> None:
+    """Configure filter thresholds."""
+    apply_startup(ctx)
+    run_async(
+        filter_criteria_impl(
+            ctx.obj.config,
+            min_uniqueness=min_uniqueness,
+            min_sub_ratio=min_sub_ratio,
+            max_cross_dupe=max_cross_dupe,
+            min_cyrillic=min_cyrillic,
+        )
+    )
+
+
+@settings_app.command("reactions")
+def settings_reactions(
+    ctx: typer.Context,
+    min_interval: int | None = typer.Option(
+        None,
+        "--min-interval",
+        help="Minimum seconds between reactions per account (clamped to 1–300; default 30)",
+    ),
+) -> None:
+    """Configure reaction sending cadence."""
+    apply_startup(ctx)
+    run_async(reactions_impl(ctx.obj.config, min_interval=min_interval))
+
+
+@settings_app.command("semantic")
+def settings_semantic(
+    ctx: typer.Context,
+    provider: str | None = typer.Option(None, "--provider", help="Embedding provider"),
+    model: str | None = typer.Option(None, "--model", help="Embedding model"),
+    api_key: str | None = typer.Option(None, "--api-key", help="Embedding API key"),
+) -> None:
+    """Configure semantic search."""
+    apply_startup(ctx)
+    run_async(
+        semantic_impl(
+            ctx.obj.config, provider=provider, model=model, api_key=api_key
+        )
+    )
